CrossLanguageEmpirical/requirements.txt
google-cloud-bigquery==1.8.0
pandas==0.25.1
